Created separate files for gardient container and styled text
