+++ lib/gradient_container.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_app/styled_text.dart';
+
+class GradientContainer extends StatelessWidget {
+  const GradientContainer({super.key});
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: const BoxDecoration(
+        gradient: LinearGradient(
+          colors: [
+            Colors.teal,
+            Colors.blueAccent,
+            Colors.cyanAccent,
+          ],
+          begin: Alignment.topLeft,
+          end: Alignment.bottomRight,
+        ),
+      ),
+      child: const Center(
+        child: StyledText(),
+      ),
+    );
+  }
+}
